COZMO-17622 Created current and deprecated tweepy wrapper (#53)
* updated twitter wrappers

* updated desktop security guard to not crash if faces disappear while cozmo is in a process
